+++ packages/showcases/css/stories/components/modal/examples/demo.runscript.js
@@ -1,12 +1,18 @@
 window.addEventListener('DOMContentLoaded', (event) => {
   const templateModal = `
-    <div class="vtmn-modal" id="vtmn-modal">
+    <div class="vtmn-modal" id="vtmn-modal" aria-hidden="true">
+      <div id="vtmn-modal-background" class="vtmn-modal_background-overlay"></div>
       <div class="vtmn-modal_content">
         <div class="vtmn-modal_content_title">
-          <p class="vtmn-modal_content_title--text">
-            What is a modal?
-          </p>
-          <button id="btn-close-modal-1">
+          <p class="vtmn-modal_content_title--text">What is a modal?</p>
+          <button
+            id="btn-close-modal-1"
+            class="
+              vtmn-btn
+              vtmn-btn_variant--ghost
+              vtmn-btn_size--small
+              vtmn-btn--icon-alone" 
+              aria-label="close">
             <span class="vtmx-close-line"></span>
           </button>
         </div>
@@ -16,19 +22,33 @@ window.addEventListener('DOMContentLoaded', (event) => {
             assumenda? Asperiores rem nulla odit saepe dolores molestias
             exercitationem accusamus perferendis est aut repudiandae optio vel dicta
             reprehenderit ad, repellendus officiis cumque omnis labore in quia?
+            Lorem ipsum dolor, sit amet consectetur adipisicing elit. Numquam,
+            assumenda? Asperiores rem nulla odit saepe dolores molestias
+            exercitationem accusamus perferendis est aut repudiandae optio vel dicta
+            reprehenderit ad, repellendus officiis cumque omnis labore in quia?
+            Lorem ipsum dolor, sit amet consectetur adipisicing elit. Numquam,
+            assumenda? Asperiores rem nulla odit saepe dolores molestias
+            exercitationem accusamus perferendis est aut repudiandae optio vel dicta
+            reprehenderit ad, repellendus officiis cumque omnis labore in quia?
+            Lorem ipsum dolor, sit amet consectetur adipisicing elit. Numquam,
+            assumenda? Asperiores rem nulla odit saepe dolores molestias
+            exercitationem accusamus perferendis est aut repudiandae optio vel dicta
+            reprehenderit ad, repellendus officiis cumque omnis labore in quia?
+            Lorem ipsum dolor, sit amet consectetur adipisicing elit. Numquam,
+            assumenda? Asperiores rem nulla odit saepe dolores molestias
+            exercitationem accusamus perferendis est aut repudiandae optio vel dicta
+            reprehenderit ad, repellendus officiis cumque omnis labore in quia?
           </p>
-        <div class="vtmn-modal_content_body--overflow-indicator"></div>
+          <div class="vtmn-modal_content_body--overflow-indicator"></div>
+        </div>
+        <div class="vtmn-modal_content_actions">
+          <button id="btn-close-modal-2" class="vtmn-btn vtmn-btn_variant--secondary">Got Back</button>
+          <button id="btn-close-modal-3" class="vtmn-btn vtmn-btn_variant--primary">
+            Yes, I understand
+          </button>
+        </div>
       </div>
-      <div class="vtmn-modal_content_actions">
-        <button id="btn-close-modal-2" class="vtmn-btn vtmn-btn_variant--secondary">
-          Got Back
-        </button>
-        <button id="btn-close-modal-3" class="vtmn-btn vtmn-btn_variant--primary">
-          Yes, I understand
-        </button>
-      </div>
-    </div>
-  </div>`;
+    </div>`;
 
   function createModal() {
     //Create the block div container
@@ -58,6 +78,13 @@ window.addEventListener('DOMContentLoaded', (event) => {
 
     document
       .getElementById(`btn-close-modal-3`)
+      .addEventListener('click', function () {
+        document.getElementById(`vtmn-modal`).remove();
+        document.getElementById('btn-modal').disabled = false;
+      });
+
+    document
+      .getElementById(`vtmn-modal-background`)
       .addEventListener('click', function () {
         document.getElementById(`vtmn-modal`).remove();
         document.getElementById('btn-modal').disabled = false;